RDA5807 Arduino Library

+++ src/RDA5807.h
@@ -465,32 +465,32 @@ typedef union
 typedef union {
     struct
     {
-        uint8_t address : 2;            // Depends on Group Type and Version codes. If 0A or 0B it is the Text Segment Address.
-        uint8_t DI : 1;                 // Decoder Controll bit
-        uint8_t MS : 1;                 // Music/Speech
-        uint8_t TA : 1;                 // Traffic Announcement
-        uint8_t programType : 5;        // PTY (Program Type) code
-        uint8_t trafficProgramCode : 1; // (TP) => 0 = No Traffic Alerts; 1 = Station gives Traffic Alerts
-        uint8_t versionCode : 1;        // (B0) => 0=A; 1=B
-        uint8_t groupType : 4;          // Group Type code.
+        uint16_t address : 2;            // Depends on Group Type and Version codes. If 0A or 0B it is the Text Segment Address.
+        uint16_t DI : 1;                 // Decoder Controll bit
+        uint16_t MS : 1;                 // Music/Speech
+        uint16_t TA : 1;                 // Traffic Announcement
+        uint16_t programType : 5;        // PTY (Program Type) code
+        uint16_t trafficProgramCode : 1; // (TP) => 0 = No Traffic Alerts; 1 = Station gives Traffic Alerts
+        uint16_t versionCode : 1;        // (B0) => 0=A; 1=B
+        uint16_t groupType : 4;          // Group Type code.
     } group0;
     struct
     {
-        uint8_t address : 4;            // Depends on Group Type and Version codes. If 2A or 2B it is the Text Segment Address.
-        uint8_t textABFlag : 1;         // Do something if it chanhes from binary "0" to binary "1" or vice-versa
-        uint8_t programType : 5;        // PTY (Program Type) code
-        uint8_t trafficProgramCode : 1; // (TP) => 0 = No Traffic Alerts; 1 = Station gives Traffic Alerts
-        uint8_t versionCode : 1;        // (B0) => 0=A; 1=B
-        uint8_t groupType : 4;          // Group Type code.
+        uint16_t address : 4;            // Depends on Group Type and Version codes. If 2A or 2B it is the Text Segment Address.
+        uint16_t textABFlag : 1;         // Do something if it chanhes from binary "0" to binary "1" or vice-versa
+        uint16_t programType : 5;        // PTY (Program Type) code
+        uint16_t trafficProgramCode : 1; // (TP) => 0 = No Traffic Alerts; 1 = Station gives Traffic Alerts
+        uint16_t versionCode : 1;        // (B0) => 0=A; 1=B
+        uint16_t groupType : 4;          // Group Type code.
     } group2;
     struct
     {
-        uint8_t content : 4;            // Depends on Group Type and Version codes.
-        uint8_t textABFlag : 1;         // Do something if it chanhes from binary "0" to binary "1" or vice-versa
-        uint8_t programType : 5;        // PTY (Program Type) code
-        uint8_t trafficProgramCode : 1; // (TP) => 0 = No Traffic Alerts; 1 = Station gives Traffic Alerts
-        uint8_t versionCode : 1;        // (B0) => 0=A; 1=B
-        uint8_t groupType : 4;          // Group Type code.
+        uint16_t content : 4;            // Depends on Group Type and Version codes.
+        uint16_t textABFlag : 1;         // Do something if it chanhes from binary "0" to binary "1" or vice-versa
+        uint16_t programType : 5;        // PTY (Program Type) code
+        uint16_t trafficProgramCode : 1; // (TP) => 0 = No Traffic Alerts; 1 = Station gives Traffic Alerts
+        uint16_t versionCode : 1;        // (B0) => 0=A; 1=B
+        uint16_t groupType : 4;          // Group Type code.
     } refined;
     uint16_t blockB;
 } rds_blockb;
